requirements.txt
streamlit>=1.40.0
mcp>=1.20.0
groq>=0.9.0
chromadb>=0.5.3
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.108.0
python-telegram-bot>=20.7
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
requests>=2.32.3
pydantic>=2.12.0
sqlalchemy>=2.0.23
numpy>=2.0.0
